some crash issue fixed and  decimal speed filed

+++ app/src/main/java/com/shazcom/gps/app/data/response/DeviceData.kt
@@ -230,7 +230,7 @@ data class Items(
     val lat: Double?,
     val lng: Double?,
     val course: Int?,
-    val speed: Int?,
+    val speed: Double?,
     val altitude: Int?,
     val icon_type: String?,
     val icon_color: String?,
